tests: handling of default values in leafs
RESTCONF RFC [1] says that server should report the mode of handling
default values. We have not implemented capabilities list yet but we
should choose how the default values should be handled. This affects GET
operations that do not specify with-defaults query parameter, and create
operations on nodes with default values. These modes are defined in a
separate RFC that deals with this problem in NETCONF [2].

We rely on libyang+sysrepo for the implementation. By default they work
with explicit default data mode. This commit adds some test to check for
that.

[1] https://datatracker.ietf.org/doc/html/rfc8040#section-3.5.4
[2] https://datatracker.ietf.org/doc/html/rfc6243#section-2

Change-Id: I5548f65463b02c8919323e12b6bfd36d2b321526