restconf: fix error-type and tag for DELETE operation
The RFC [1] is not very specific in what error report when user tries
to delete a node using DELETE method. There is some discussion
regarding that in the NETCONF mailing list [2].

Originally we used HTTP 404 and error-tag invalid-value because I
probably misunderstood the wording in the second paragraph of the RFC.
My current work on implementing default-handling basic-mode revealed
this nonsense.

After reading through the mail thread we have decided to go with HTTP
404 and error-tag data-missing (as in the NETCONF edit-config) as
proposed by [3].

We are unsure about that one test with FIXME comment. There is some
discussion in sysrepo's bugtracker [4].

[1] https://datatracker.ietf.org/doc/html/rfc8040#section-4.7
[2] https://mailarchive.ietf.org/arch/msg/netconf/XcF9r3ek3LvZ4DjF-7_B8kxuiwA/
[3] https://mailarchive.ietf.org/arch/msg/netconf/hkVDdHK4xA74NgvXzWP0zObMiyY/
[4] https://github.com/sysrepo/sysrepo/issues/3283

Change-Id: I1082e3abf99948b0ad22677b71d8d8a319bc2786