restconf: implement insert=before/after query parameters
A RESTCONF server must support inserting list entries list user ordered
lists using insert and point query parameters [1,2].

Unfortunately, the point query parameter, which specifies the insertion
point, contains a full path to the list entry (not only keys), which
makes the implementation a little bit more difficult than if only the
keys were specified.
The problem is that when using the POST method, we do not know what is
the modified list until we read the payload. Therefore, we can't
validate the point URI validity in the parser itself.

Fortunately, the insertion logic itself can be offloaded to
libyang+sysrepo.

[1] https://datatracker.ietf.org/doc/html/rfc8040#section-4.8.5
[2] https://datatracker.ietf.org/doc/html/rfc8040#section-4.8.6

Change-Id: I88522e8c2838688078bf899a6982f0af6b590e73
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/dependencies/+/7114