restconf: fix location for ietf-yang-library submodules
There was a logic error in rewriting a location leaf name in
ietf-yang-library output for submodules inside module-set list.

Such modules are nested one level below the ordinary modules. The code
checking if we are inside module-set list or modules-state container did
not expect that. This resulted in locationLeafName variable set to empty
string.

Fixes: 9ff26d45b9c6535017c673b9aff0a57a64142c8c
Change-Id: Ic3555866d8d32f55f184d1de0c87dcd5339e0323