restconf: output location leaf only if sysrepo reports it
There was a bug in the code adding the location leafs. We added them to
every module even if sysrepo did not return the location. This could
happen in situations where NACM rules set the location as not
readable.

This patch fixes the behaviour by adding the location leaf for a module
only if the location was reported by sysrepo.

Fixes: 9ff26d45b9c6535017c673b9aff0a57a64142c8c
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/dependencies/+/7190
Change-Id: I55a037dd7bfd3965d4cf51c4484fe8a24835ebdf