restconf: partialy fix data race in PUT method
PUT returns HTTP status codes, one if the node was created and a
different one when node was modified. The code implements this using a
query (that is executed before the actual edit) to check if the node was
there before the edit.

This is potentially racy, because the node can be created between the
query and the actual edit.

This patch adds a sysrepo lock so this does not happen. There is one
problem, though. The candidate datastore reset itself to mirror running
when sr_unlock is called [1].
Therefore we do not lock the candidate DS.

[1] https://github.com/sysrepo/sysrepo/blob/4f40f5406f24e5a565d0ee554389fff2aad4bd1b/doc/main.dox#L94

Bug: https://github.com/CESNET/rousette/issues/2
Change-Id: Ie211018e921b5b2d7a2f78d8cc531d58db6ac167