restconf: publish stream list
There is a container in the ietf-restconf-monitoring module that is used
to publish stream resources locations for users [1].

This patch populates the contents of the container when the RESTCONF
server starts.

However, there is the same problem as in 9ff26d4 ("restconf: show
schema locations for modules in ietf-yang-library"). We have to provide
full URLs with server name, i.e., we have to parse it from the Host HTTP
header and then edit the YANG data tree.

[1] https://datatracker.ietf.org/doc/html/rfc8040#section-6.2

Change-Id: Ibd2250c9b7033e47356c1b2c08dff3dce78e2061