EventStream: move signal connect to constructor
In the future code for notifications I require to activate the signal
earlier. (When I construct the notification callbacks, the signal must
already be connected. But also, the server must not yet respond with
res.end() because things might go wrong. (See the next patches)).

So, after some brainstorming time we decided to move the signal connect
code to the EventStream constructor, as well as the initialEvent push.

We also discussed the possibility of removing the initialEvent
parameter in favour of manually invoking the signal between EventStream
construction and calling the activate() method. However, this is not
possible, because there is only one optical event data provider and
manually invoking the signal might send the same data twice to clients,
that were already listening.

Change-Id: I45d951206479ec7b0b2aff601bd3379744ee3e17