restconf: implement HEAD HTTP method
The HTTP HEAD [1] method behaves the same as GET method, but it does
not contain the response body.
Fortunately, nghttp2 does not send the response body in this case, so
the implementation does not need to be changed much.

We add support for using HEAD method also for /streams and /yang
resources.

[1] https://datatracker.ietf.org/doc/html/rfc8040#section-4.2

Change-Id: Ic10c4707ce2cfb25d7f0e416b5b292d81802f205