restconf: omit HTTP method from NETCONF error descriptions
The error descriptions returned from the PUT method are repeating
what is already stated in the error-tag and also they say that the error
happened in the PUT HTTP method. I think that it is redundant.

Also, it helps for the next commits where we reuse the processPut
function for HTTP PATCH as well.

Change-Id: I06aecac5d27d944fb87f20a000df11b7f34185ee