restconf: refactor constructing HTTP headers for OPTIONS
In the two commits before

 * 0d2f394 ("restconf: implement OPTIONS HTTP method")
 * 7e84a59 ("restconf: add allow header when returning 405")

we added a code for creating the content for the allow HTTP header which
is required for OPTIONS responses and for responses with 405 status code.
This will need to be expanded a when the PATCH method is implemented
because more headers will be added to such responses. This patch allows
us to create only a small diff in the future.

Change-Id: I3e5755a245b789823c3eafed0a1c4ce4052ff570