restconf: refactor rejectWithError
We are working on YANG patch support for rousette. In the YANG patch
code we will want to use rejectWithError too, but the error output has a
similar schema but it is nested in some other nodes [1].

This commits moves the common code from rejectWithError into a separate
function that does not create nodes directly under /ietf-restconf:errors
but rather under a node that is provided.

[1] https://datatracker.ietf.org/doc/html/rfc8072#section-2.3

Change-Id: I68fb19a3b00b85de7a3807171913584f7d2a71db