tests: notifications: propagate exceptions to the main thread
When the "background" thread (which just happens to check all the
expectations for being fulfilled, see the previous commit) throws an
exception, that leads directly to std::terminate, which is Not Nice™,
and it also happens to break the output of doctest in a rather
spectacular way, especially under ASAN/UBSAN.

Let's just propagate these exceptions to the main thread. It seems that
no information is lost, and that trompeloeil will still happily print
out which expectations were left unfulfilled.

Change-Id: I70179156676c4dcb04029cb69c8b79682f207f6c