restconf: refactor URI parser grammar rules
While working on adding something to URI grammar I found the grammar a
bit odd and maybe a little "hacky" because the yang-library-version
branch was imlemented as an extra rule. This is not necessary, it is
just another resource which, however, is not followed by uriPath rule.

This led me to think that we can cleanup a little bit and merge some
rules.

Change-Id: I2d4b2e8201d9fdd0802bdfadfc6e316de5b27df3