tests: notifications: prevent UB in thread synchronization
It's always an error to call count_down() twice on an std::latch with a
capacity of one. Because the intention here is clearly to make sure that
some events are only sent *before* the client has started and some
others sent *after* the client has connected, let's use a pair of
latches for that.

Fixes: f5bdf83 restconf: support for start-time and stop-time parameters in streams
Change-Id: If5453b620dc82c10a2f7d198d0cf6c6f86cb7339