restconf: support fully qualified identityrefs in URI list keys
Our URI parser was unable to parse list key values which were fully
qualified identifiers. This could happen when the list key's type was
a identityref.

The ABNF in RFC 8040 [1] specifies that the key-value is a string and we
implemented it as a string. However, it also specifies that the
key-value conforms to RFC 7951's JSON identifier encoding rules, which
we missed in the initial implementation. This patch fixes that.

[1] https://datatracker.ietf.org/doc/html/rfc8040#section-3.5.3.1
[2] https://datatracker.ietf.org/doc/html/rfc7951#section-4

Bug: https://github.com/CESNET/rousette/issues/12
Fixes: e06d5bff672df7d6942ab984a1c455c77b95f1d9 ("server: parser for data resources in URI paths")
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/dependencies/+/7729
Change-Id: I44245d831e8de6d0e6f991fcd18319c095b49b1d