restconf: use parser on the raw path to preserve percent encoding
The difference between nghttp2's uri_ref::path and raw_path is that the
raw_path keeps the percent encoding, while the path converts the percent
encoded chars to their normal form.

Our parser expects some parts of the URI to be percent encoded so we
have to use raw_paths everywhere.

I thought about rewriting the parser not to expect percent encoded
characters but that would bring some complications. For instance when
querying lists, the RESTCONF RFC specifies that every key value is
percent encoded and individual key values are delimited by commas [1].
So, when somebody sends a request like /restconf/data/list=a%2Cb,c the
normal form is /restconf/data/list=a,b,c and in that case we obtain
three keys, but the client send only two, where the first one contained
comma.

Change-Id: I473501cef3c8eae9af0c5d0751393cdad647e23c