tests: move stuff from the header file into cpp file
Allright, I have had enough. I am no longer waiting for a minute for a
recompilation of one test.
This patch splits the aux-utils.h file into a cpp file and two headers.
One of the headers is only for declarations of the functions and
datatypes in the cpp file, the second header implements all the helper
functions that require SERVER_PORT and further helper constants for the
restconf tests.

Compile times are a little better. I have measured compilation times of
two arbitrary restconf tests with asan+ubsan (and ccache disabled):

 - restconf-reading: 17.05s -> 13.85s
 - restconf-delete: 15.18s -> 11.87s

Not ideal, but it is certainly better.

Change-Id: If529cbc8954d50494711a408231ea4c2c4daf072