tests: fix deadlock in tests
In the case the client never requests (because of misconfiguration or
something else) the test would still be waiting for somebody to
decrement the latch.
Unfortunately std::latch does not allow waiting for specified amount of
time, so this patch replaces it with std::binary_semaphore which can
fail after specified amount of time.

Change-Id: I38c007555bb5ab543329d724a16f024a9d80903e