restconf: prepare support for subscribed notifications
This commit adds support for subscribed notifications in the URI parser
and server processing. We are going to implement the required RPC
handlers for creating subscribed notifications directly in rousette on
the HTTP-level, i.e., we are not going to pass those RPC calls to
sysrepo. The reason behind this is that netopeer2 server already
implements the RPC callback too and it would be impossible to run
rousette with netopeer2 server at the same time.

Change-Id: I207e4d5b0d34eea4477c50019e9cab4f8e81e1a6