restconf: prepare support for subscribed notifications
This commit adds support for establishing subscribed notifications [1,2]
in the URI parser and server processing. We are going to implement the
required RPC handlers for creating subscribed notifications directly in
rousette on the HTTP-level, i.e., we are not going to pass those RPC calls
to sysrepo.

The Netopeer2 project already implements the RPC callbacks for subscribed
notifications RPCs. Sysrepo can handle multiple RPC callbacks, but only
one of them can return data. The other will be executed but the data will
be ignored [4].
It makes no sense to establish subscriptions through NETCONF and then
request the stream via RESTCONF and AFAIK it is even not allowed by RFCs
which states that NETCONF requests are handled by the same NETCONF
session and also, there is no way how to get to those subscriptions via
RESTCONF. RESTCONF provides its own way how to read the streamed
notifications [3,4].

And yes, we want both Netopeer2 and Rousette to be able to run at the
same time on the same machine.

[1] https://datatracker.ietf.org/doc/html/rfc8639
[2] https://datatracker.ietf.org/doc/html/rfc8650
[3] https://datatracker.ietf.org/doc/html/rfc8640
[4] https://github.com/sysrepo/sysrepo/blob/ef93a1253cc97f13671759f6e7790cbf729a5ae9/doc/subs.dox#L78

Change-Id: I207e4d5b0d34eea4477c50019e9cab4f8e81e1a6