Always print empty containers in user-defined content
There's some subtle interaction between depth-limited printing, default
values, and empty non-presence containers. In order to prepare for an
upstream change in sysrepo which fixed depth-limited printing, let's
stick to "always" printing the empty non-presence containers.

(It actually isn't "always" because the data such as error reports,
patches, etc., are still printed using the default settings. But the
data which is subject to user-defined constraints is now consistent.)

Change-Id: Ie2d6ad2e851d667e7e7582fcd0ac75dabce2c05e