requests: create sysrepo session later
I was looking at all places where a sysrepo session is used, and tried
to defer their creations until the last moment possible, just to
(potentially) save some resources when it isn't needed. Some of these
changes are cosmetic, but I was touching that code anyway.

This cannot be done for the generic `restconfRoot` handler because of
its associated error handling which requires a libyang context. Well, it
*could* be done, but I don't feel like doing some m_monitoringSession
magic for little to no purpose.

Change-Id: Ia84c90abc1464e4a0f9682779d7471aefad9e2de