restconf: add {kill,delete}-subscription RPCs
Add implementations for deleting and killing subscriptions [1,2].

Rousette allows establishing subscriptions by anonymous users. Anonymous
users can not call delete-subscription RPC though because we do not know
if it is the same anonymous user who initiated the subscription.
The subscriptions are killed after a minute of inactivity, so the
subscription will get deleted by rousette automatically after the client
disconnects.

[1] https://datatracker.ietf.org/doc/html/rfc8639.html#section-2.4.4
[2] https://datatracker.ietf.org/doc/html/rfc8639.html#section-2.4.5

Change-Id: I443d6fd88f2797045b5ae0a7d7e1a761fe74e7cb