restconf: simplify condition in NACM anonymous access
The condition required that all access-operations values are "read". But
access-operations is a union, so it is enough to check that we have only
one value in the union and that it is "read".

Also test what happens if anonymous rule access-operation is a union of
read and something else.

Change-Id: I291243d9e2167601b24d49684d53c361407f66be