restconf: catch any exceptions in withRestconfExceptions
In [1] we saw a weird asio behaviour when withRestconfExceptions threw
an exception. The exception was not caught and rousette started behaving
in a strange way.

This commit wraps the withRestconfExceptions call in a try-catch block
catching *any* exception. Rousette should now not crash in such
situations and attempt to log exception's what() message if the
exception inherits from std::exception.

[1] https://github.com/CESNET/rousette/issues/19
Fixes: https://github.com/CESNET/rousette/issues/19

Change-Id: I2c090b9a76b062101ba422a7d50e8e699779e203