restconf: endpoint for subscribed notifications
This patch creates a new endpoint for the subscribed notification
streams in the RESTCONF server. The previous commit added the
establish-subscription RPC [1], which returned a stream URI, that was
not accessible in that patch. The patch was extra large even without
that functionality so I have split it into two commits.

Also, I shortened the locked section in EventStream's implementation of
nghttp's on_close callback. Without doing that, we get a lock order
inversion because:
Scenario 1: A thread acquires the mutex (let's call it M1) in
DynamicNotificationHTTPStream::awaitNextNotification and
then sends the notification signal, which acquires the mutex inside
EventStream (mutex M2).
In the other scenario, the nghttp's on_close callback acquired M2 and
then proceeds to call onClientDisconnectedCb which, in our case,
acquires M1.
Thus there is a potential for deadlock reported by TSAN.

[1] https://datatracker.ietf.org/doc/rfc8639/

Change-Id: I07dcdd3fb9ef4f05f93ae6dda0e0d71094a0bbfc