restconf: endpoint for subscribed notifications
This patch creates a new endpoint for the subscribed notification
streams in the RESTCONF server. The previous commit added the
establish-subscription RPC [1], which returned a stream URI, that was
not accessible in that patch. The patch was extra large even without
that functionality so I have split it into two commits.

[1] https://datatracker.ietf.org/doc/rfc8639/

Change-Id: I07dcdd3fb9ef4f05f93ae6dda0e0d71094a0bbfc