refactor: restconf stream request types
In commit 97ceef1 ("restconf: refactor uri parser for stream URIs") we
prepared the URI parser for the new subscribed stream notifications.
What I did not think through was that the subscribed notifications do
not have query parameters. However, I refactored the types in a way
that the query parameters member is now in the common base type.

This patch fixes that by editing the types in a way that only the
stream notifications will have the query parameters member. The
type for subscribed notifications (which will come in a followup patch)
will not have it.

See-also: 97ceef119c900c37bbaa27860c3b43cfa6d69f95 ("restconf: refactor uri parser for stream URIs")
Change-Id: Idc1e05f7ff05cea84b5806f60e217605f75e7a9f