restconf: validate input for internal rpcs
Jan found an issue where invalid input to internal rpcs was accepted. It
turns out that the issue was in parsing, but not validating input for
internally processed RPCs (those RPCs for subscribed notifications).

This patch adds the validation using a wrapper for lyd_validate_op from
libyang-cpp.

I shuffled a little bit of the code around together with this change to
make it cleaner.
Also, to make the code generic, we store the name of the internal RPC
and a XPath for fetching data required for validation (see
lyd_validate_op docs). The code validates the input data before processing
the RPC.

Change-Id: I55296895e6993b4ca27d21ce2a64ec2d159a35fc