port to libyang v4
- the prefix of the "ietf-netconf-with-defaults" module has changed,
- RPC parsing needs a flag to reject unknown input data nodes,
- due to the "printed context" mess, we now need an extra flag to ensure
  that we can still access the YANG schema source.

The last item is interesting. We could also read directly from the
on-disk location, but that looks a bit meh. It's also a bit annoying
that one has to change this thing over a lot of places, but I think that
this is much nicer than trying to force a change over an existing
connection (e.g., via a connection-specific call in the Server::Server
ctor which might hit some internal locks depending on how the sysrepo
state look like).

I was considering adding a wrapper for `ly_ctx_get_options()` and a
check in Server::Server, but then I decided that I won't bother.

Change-Id: I142f650affb53411b5c1dd81311400a2aa6dec20
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/dependencies/+/8990